Updated Registration to be less opaque by splitting the combination warp of modality to target into its own script. Unified, mostly, the naming conventions of final warps. Updated resampling to be trilinear as sometimes fsl's spline interpolation can be weird. Updated registrations to allow for dataspace final targets (useful mostly for DTI and retinotopy).

+++ SCRIPTS/Registration/DistortionCorrection_APPA.csh
@@ -22,6 +22,11 @@ set ped = ($5)
 set fm = ($6)
 set Reg_Target = $7
 
+if($Reg_Target != T1) then
+	echo "BBR cannot be performed without the target being T1."
+	exit 1
+endif
+
 set SubjectHome = $cwd
 
 rm -rf ${SubjectHome}/Anatomical/Volume/FieldMapping_${FM_Suffix}
